# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Portfolio API",
        "description": "This is the Swagger UI documentation for my Laravel API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/portfolio": {
            "get": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Get all portfolio",
                "operationId": "5845d8d04917a76915a777d8c988566f",
                "responses": {
                    "200": {
                        "description": "List of portfolio"
                    }
                }
            },
            "post": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Create a new portfolio",
                "operationId": "6240b775a71c2be391a2b9d0d3ed5ca7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "p_title",
                                    "p_category",
                                    "p_img",
                                    "p_detail",
                                    "display"
                                ],
                                "properties": {
                                    "p_img": {
                                        "type": "integer"
                                    },
                                    "p_title": {
                                        "type": "string"
                                    },
                                    "p_category": {
                                        "type": "string"
                                    },
                                    "p_detail": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Portfolio created successfully"
                    }
                }
            }
        },
        "/api/portfolio/{id}": {
            "get": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Get an portfolio by ID",
                "operationId": "eee3f303f0feef0fd119a697a4033503",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Portfolio retrieved successfully"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Update a portfolio by ID",
                "operationId": "d1ce9341849cacbeabb4aa4aa44ecfe9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "p_img": {
                                        "type": "integer"
                                    },
                                    "p_title": {
                                        "type": "string"
                                    },
                                    "p_category": {
                                        "type": "string"
                                    },
                                    "p_detail": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portfolio updated successfully"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                }
            }
        },
        "/api/portfolio/visible/{id}": {
            "put": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Delete an portfolio by ID",
                "operationId": "917b2f8d40cfc048c36dc8a7b7c14683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete portfolio successfully"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                }
            }
        },
        "/api/portfolio/reorder": {
            "put": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Reorder portfolios",
                "operationId": "ab359e4f78cb14987e4d5b18ccb45d30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "p_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "p_order": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portfolio order updated successfully"
                    }
                }
            }
        },
...
    },
...
}